Move mocks to conftest.py
