+++ tests/cli/config/conftest.py
@@ -1,7 +1,5 @@
 from pytest_cases import fixture
 
-from statue.config.configuration_builder import ConfigurationBuilder
-
 
 @fixture
 def mock_find_sources(mocker):
@@ -11,11 +9,6 @@ def mock_find_sources(mocker):
 @fixture
 def mock_expend(mocker):
     return mocker.patch("statue.cli.config.expend")
-
-
-@fixture
-def mock_configuration_path(mocker):
-    return mocker.patch.object(ConfigurationBuilder, "configuration_path")
 
 
 @fixture
